backend/requirements.txt
asgiref==3.11.1
contourpy==1.3.2
cycler==0.12.1
distlib==0.4.0
Django==6.0.2
django-cors-headers==4.9.0
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
mysqlclient==2.2.7
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
virtualenv==20.36.1
virtualenvwrapper-win==1.2.7
